RAGChat/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
backoff==2.2.1
bcrypt==5.0.0
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.0.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.3.1
coloredlogs==15.0.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.0.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
Django==6.0.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
durationpy==0.10
fastapi==0.128.0
filelock==3.20.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.12.0
google-auth==2.47.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Incremental==24.11.0
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==34.1.0
langchain==0.0.340
langsmith==0.0.92
markdown-it-py==4.0.0
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==0.28.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
overrides==7.7.0
packaging==25.0
posthog==7.5.1
propcache==0.4.1
protobuf==6.33.2
pulsar-client==3.9.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pypdf==3.16.3
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.7.0
social-auth-core==4.8.3
SQLAlchemy==2.0.45
sqlparse==0.5.5
starlette==0.50.0
sympy==1.14.0
tenacity==8.5.0
tokenizers==0.22.2
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.3.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
zope.interface==8.1.1
